request controller fixes
