--- BikeRent.Server/BikeRent.Server.sln
@@ -1,25 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 17
-VisualStudioVersion = 17.11.35312.102
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "BikeRent.Server", "BikeRent.Server.csproj", "{AE3D18B5-AC54-4651-AF0D-A9A3ECC64EC4}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{AE3D18B5-AC54-4651-AF0D-A9A3ECC64EC4}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{AE3D18B5-AC54-4651-AF0D-A9A3ECC64EC4}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{AE3D18B5-AC54-4651-AF0D-A9A3ECC64EC4}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{AE3D18B5-AC54-4651-AF0D-A9A3ECC64EC4}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {84C1A816-FAC8-4CF1-ABE3-FB216CE89D26}
-	EndGlobalSection
-EndGlobal
