+++ Solution BikeRent.sln
@@ -7,7 +7,7 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "BikeRent.Domain", "BikeRent
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "BikeRent.Tests", "BikeRent.Tests\BikeRent.Tests.csproj", "{EF560E16-C60D-436B-80CD-2FC8E57C5473}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "BikeRent.Server", "Server\BikeRent.Server.csproj", "{007A2F15-B9C9-4A09-AFA5-F67F014FD353}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "BikeRent.Server", "BikeRent.Server\BikeRent.Server.csproj", "{5CE25FDE-F2CD-40BF-894D-35345969245C}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -23,10 +23,10 @@ Global
 		{EF560E16-C60D-436B-80CD-2FC8E57C5473}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{EF560E16-C60D-436B-80CD-2FC8E57C5473}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{EF560E16-C60D-436B-80CD-2FC8E57C5473}.Release|Any CPU.Build.0 = Release|Any CPU
-		{007A2F15-B9C9-4A09-AFA5-F67F014FD353}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{007A2F15-B9C9-4A09-AFA5-F67F014FD353}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{007A2F15-B9C9-4A09-AFA5-F67F014FD353}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{007A2F15-B9C9-4A09-AFA5-F67F014FD353}.Release|Any CPU.Build.0 = Release|Any CPU
+		{5CE25FDE-F2CD-40BF-894D-35345969245C}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{5CE25FDE-F2CD-40BF-894D-35345969245C}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{5CE25FDE-F2CD-40BF-894D-35345969245C}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{5CE25FDE-F2CD-40BF-894D-35345969245C}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
